backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
bcrypt==3.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.2
flatbuffers==25.12.19
fonttools==4.62.1
fpdf2==2.8.7
gast==0.7.0
google-ai-generativelanguage
google-api-core
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos
greenlet==3.3.2
groq==1.1.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
keras==3.13.2
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
mysql-connector-python==9.6.0
namex==0.1.0
numpy==2.4.3
openai==2.29.0
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
passlib==1.7.4
pillow==12.1.1
proto-plus==1.27.1
protobuf
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rich==14.3.3
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==1.0.0
tensorflow-cpu
termcolor==3.3.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
wheel==0.46.3
wrapt==2.1.2
psycopg2-binary==2.9.10
